Refactor to ensure SyncTransferContrller not adding additional dependency.
